Keep exemptions in case of closing a tab and subsequently undoing the action (#4905)
Task/Issue URL:
https://app.asana.com/0/488551667048375/1207991167374138/f

### Description
Updated the code to make sure the exemption is kept after undoing a tab
close action.

### Steps to test this PR
See the **Tests** section in the description of
https://app.asana.com/0/488551667048375/1207991167374138/f

### NO UI changes

+++ ad-click/ad-click-impl/src/main/java/com/duckduckgo/adclick/impl/DuckDuckGoAdClickData.kt
@@ -16,6 +16,7 @@
 
 package com.duckduckgo.adclick.impl
 
+import com.duckduckgo.adclick.impl.remoteconfig.AdClickAttributionFeature
 import com.duckduckgo.adclick.impl.store.exemptions.AdClickExemptionsDao
 import com.duckduckgo.adclick.impl.store.exemptions.AdClickExemptionsDatabase
 import com.duckduckgo.adclick.impl.store.exemptions.AdClickTabExemptionEntity
@@ -50,6 +51,7 @@ class DuckDuckGoAdClickData(
     val database: AdClickExemptionsDatabase,
     private val coroutineScope: CoroutineScope,
     private val dispatcherProvider: DispatcherProvider,
+    private val adClickAttributionFeature: AdClickAttributionFeature,
     isMainProcess: Boolean,
 ) : AdClickData {
 
@@ -62,7 +64,7 @@ class DuckDuckGoAdClickData(
 
     init {
         coroutineScope.launch(dispatcherProvider.io()) {
-            if (isMainProcess) {
+            if (isMainProcess && adClickAttributionFeature.persistExemptions().isEnabled()) {
                 loadToMemory()
             }
         }
@@ -100,7 +102,9 @@ class DuckDuckGoAdClickData(
     override fun removeExemption() {
         tabExemptions.remove(activeTabId)
         coroutineScope.launch(dispatcherProvider.io()) {
-            adClickExemptionsDao.deleteTabExemption(activeTabId)
+            if (adClickAttributionFeature.persistExemptions().isEnabled()) {
+                adClickExemptionsDao.deleteTabExemption(activeTabId)
+            }
         }
         Timber.d("Removed exemption for active tab $activeTabId. Tab exemptions: $tabExemptions")
     }
@@ -108,15 +112,17 @@ class DuckDuckGoAdClickData(
     override fun addExemption(exemption: Exemption) {
         tabExemptions[activeTabId] = exemption
         coroutineScope.launch(dispatcherProvider.io()) {
-            adClickExemptionsDao.insertTabExemption(
-                AdClickTabExemptionEntity(
-                    tabId = activeTabId,
-                    hostTldPlusOne = exemption.hostTldPlusOne,
-                    navigationExemptionDeadline = exemption.navigationExemptionDeadline,
-                    exemptionDeadline = exemption.exemptionDeadline,
-                    adClickActivePixelFired = exemption.adClickActivePixelFired,
-                ),
-            )
+            if (adClickAttributionFeature.persistExemptions().isEnabled()) {
+                adClickExemptionsDao.insertTabExemption(
+                    AdClickTabExemptionEntity(
+                        tabId = activeTabId,
+                        hostTldPlusOne = exemption.hostTldPlusOne,
+                        navigationExemptionDeadline = exemption.navigationExemptionDeadline,
+                        exemptionDeadline = exemption.exemptionDeadline,
+                        adClickActivePixelFired = exemption.adClickActivePixelFired,
+                    ),
+                )
+            }
         }
         Timber.d("Added exemption for active tab $activeTabId. Tab exemptions: $tabExemptions")
     }
@@ -124,15 +130,17 @@ class DuckDuckGoAdClickData(
     override fun addExemption(tabId: String, exemption: Exemption) {
         tabExemptions[tabId] = exemption
         coroutineScope.launch(dispatcherProvider.io()) {
-            adClickExemptionsDao.insertTabExemption(
-                AdClickTabExemptionEntity(
-                    tabId = tabId,
-                    hostTldPlusOne = exemption.hostTldPlusOne,
-                    navigationExemptionDeadline = exemption.navigationExemptionDeadline,
-                    exemptionDeadline = exemption.exemptionDeadline,
-                    adClickActivePixelFired = exemption.adClickActivePixelFired,
-                ),
-            )
+            if (adClickAttributionFeature.persistExemptions().isEnabled()) {
+                adClickExemptionsDao.insertTabExemption(
+                    AdClickTabExemptionEntity(
+                        tabId = tabId,
+                        hostTldPlusOne = exemption.hostTldPlusOne,
+                        navigationExemptionDeadline = exemption.navigationExemptionDeadline,
+                        exemptionDeadline = exemption.exemptionDeadline,
+                        adClickActivePixelFired = exemption.adClickActivePixelFired,
+                    ),
+                )
+            }
         }
         Timber.d("Added exemption for tab $tabId. Tab exemptions: $tabExemptions")
     }
@@ -155,7 +163,9 @@ class DuckDuckGoAdClickData(
         tabAdDomains.remove(tabId)
         tabExemptions.remove(tabId)
         coroutineScope.launch(dispatcherProvider.io()) {
-            adClickExemptionsDao.deleteTabExemption(tabId)
+            if (adClickAttributionFeature.persistExemptions().isEnabled()) {
+                adClickExemptionsDao.deleteTabExemption(tabId)
+            }
         }
         Timber.d("Removed data for tab $tabId. Tab ad domains: $tabAdDomains. Tab exemptions: $tabExemptions")
     }
@@ -164,7 +174,9 @@ class DuckDuckGoAdClickData(
         tabAdDomains.clear()
         tabExemptions.clear()
         coroutineScope.launch(dispatcherProvider.io()) {
-            adClickExemptionsDao.deleteAllTabExemptions()
+            if (adClickAttributionFeature.persistExemptions().isEnabled()) {
+                adClickExemptionsDao.deleteAllTabExemptions()
+            }
         }
         Timber.d("Removed all data. Ad clicked map is empty ${tabAdDomains.isEmpty()}. Empty tab exemptions? ${tabExemptions.isEmpty()}")
     }
@@ -179,7 +191,9 @@ class DuckDuckGoAdClickData(
             }
         }
         coroutineScope.launch(dispatcherProvider.io()) {
-            adClickExemptionsDao.deleteAllExpiredTabExemptions(currentTime)
+            if (adClickAttributionFeature.persistExemptions().isEnabled()) {
+                adClickExemptionsDao.deleteAllExpiredTabExemptions(currentTime)
+            }
         }
         Timber.d("Removed all expired data. Tab exemptions: $tabExemptions")
     }
